Default settings modified

+++ Commands/Receiver/datarate.cpp
@@ -15,14 +15,14 @@ void DataRate::setDataRate(const int32_t dataRate)
 {
    if (337 <= dataRate && dataRate <= 115200)
    {
-      int8_t value =  round(10000000 / 29 / 1 / dataRate - 1);
+      uint8_t value =  round((double) 10000000 / 29 / 1 / dataRate) - 1;
       if (value <= 127)
       {
          setValue(0, 7, value);
       }
       else
       {
-         value =  round(10000000 / 29 / 8 / dataRate - 1);
+         value =  round((double) 10000000 / 29 / 8 / dataRate) - 1;
          if (value <= 127)
          {
             setValue(0, 7, value);
